Allow to configure render pipeline settings
This patch adds the possibility to configure the render pipeline.
Currently, the default render pipeline has a setting to enable the debug
grid.

In the future several settings could be added. For now these settings
are only used at the initialization stage of the render pipeline. It
might be useful to change the render pipeline settings during the
execution.