Add tubereng-gltf crate
As often, inspired by bevy, I believe that supporting gltf format is a
must have for the game engine as it provides a way to pack entire scenes
with rigged models and such.

I'm not sure yet how it will integrate with the engine. I might have do
something similar to what bevy does on that aspect.