graphics: Fix grid shader compilation failure
This patch fixes a shader compilation error that occured on certain
platforms (i.e. intel/mesa).

fwidth() was called outside of the entry point of the fragment shader.
During the vertex shader compilation stage, the function calling
fwidth() was not culled, causing a compilation error.
(See gfx-rs/naga#2524)

As a temporary fix, the call to fwidth() has been moved to the fragment
shader entry point.